♻️ Refactor : 불필요한 훅 제거 및 적절한 컴포넌트 위치에 맞도록 로직 수정
재사용성이 없는 훅 제거 컴포넌트 위치에 맞도록 로직 수정
