+++ src/components/TravelInformationContent/ProductDescription/index.tsx
@@ -1,98 +1,74 @@
-import { useEffect, useState } from 'react';
-
-import toast from 'react-hot-toast';
-
 import ProductMap from './ProductMap';
 import CurrentLocation from './CurrentLocation';
 import CustomLocation from './CustomLocation';
 import LocationDescription from './LoactionDescription';
 
-import useDebounce from '@/src/hooks/useDebounce';
 import convertTime from '@/src/utils/convertTime';
-import instance from '@/src/api/axios';
-import { BASED_URL, TOAST_MESSAGE } from '@/src/constants/constants';
-import useMyGeolocation from '@/src/hooks/useMyGeolocation';
-import useDestinationDirection from '@/src/hooks/useDestinationDirection';
 import { placeData } from './mock';
+import { useEffect, useState } from 'react';
+import useGeolocation from 'react-hook-geolocation';
+import extractPath from '@/src/utils/extractPath';
+import { BASED_URL } from '@/src/constants/constants';
+import instance from '@/src/api/axios';
 
 const ProductDescription = () => {
-  const [showMessage, setShowMessage] = useState(false);
-  const [validKeyword, setValidKeyword] = useState(true);
-  const [customLocation, setCustomLocation] = useState('');
-  const [relatedLocation, setRelatedLocation] = useState([]);
-  const [searchCoordinate, setSearchCoordinate] = useState({ lat: 0, lng: 0 });
+  const [duration, setDuration] = useState(0);
+  const [isLoadingDirection, setIsLoadingDirection] = useState(true);
+  const [hasCurrentLocation, setHasCurrentLocation] = useState(true);
+  const [polylinePath, setPolylinePath] = useState<{ lat: number; lng: number }[]>([]);
+  const [startPoint, setStartPoint] = useState<{ lat: number; lng: number }>({ lat: 0, lng: 0 });
 
-  const { startPoint, setStartPoint, hasCurrentLocation } = useMyGeolocation();
-  const { polylinePath, duration } = useDestinationDirection(startPoint, placeData.position);
-
-  const debounceCustomLocation = useDebounce(customLocation, 500);
-
-  const getMyCoordinate = async (address: string) => {
-    const encodedAddress = encodeURIComponent(address);
-    try {
-      const response = await instance.get(
-        `${BASED_URL.KAKAO_ROAD}/local/search/address.json?analyze_type=similar&query=${encodedAddress}`,
-        { headers: { Authorization: `KakaoAK ${process.env.NEXT_PUBLIC_CLIENT_ID_KAKAO_REST}` } },
-      );
-      const hasCoordinate = response.data.documents;
-      const searchLoaction = response.data.documents[0];
-
-      if (hasCoordinate.length === 0) {
-        setValidKeyword(false);
-      }
-
-      if (hasCoordinate.length !== 0) {
-        setSearchCoordinate({ lat: parseFloat(searchLoaction.y), lng: parseFloat(searchLoaction.x) });
-        setRelatedLocation(hasCoordinate.map((location) => location.address_name));
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  const currentLocation = useGeolocation();
+  const accessLocation = currentLocation.error === null;
 
   useEffect(() => {
-    const hasCustomLocation = debounceCustomLocation.trim() !== '';
-    if (hasCustomLocation) {
-      getMyCoordinate(debounceCustomLocation);
+    const position = { lat: currentLocation.latitude, lng: currentLocation.longitude };
+    if (accessLocation) {
+      setHasCurrentLocation(true);
+      setStartPoint(position);
     }
-    if (!hasCustomLocation) {
-      setRelatedLocation([]);
+    if (!accessLocation) {
+      setHasCurrentLocation(false);
     }
-  }, [debounceCustomLocation]);
+  }, [accessLocation, currentLocation.latitude, currentLocation.longitude]);
 
-  const handleStartingPoint = () => {
-    const hasCustomLocation = customLocation.trim() !== '';
-    if (!hasCustomLocation) {
-      toast.error(TOAST_MESSAGE.EMPTY_LOCATION);
-    }
+  const isValidCoordinate = startPoint && startPoint.lat !== 0 && startPoint.lng !== 0;
 
-    if (hasCustomLocation) {
-      if (searchCoordinate === null) {
-        setValidKeyword(false);
+  useEffect(() => {
+    const getDestinationDirection = async () => {
+      const hasStartPoint = startPoint && startPoint.lat !== null && startPoint.lng !== null;
+      if (hasStartPoint) {
+        const queryParams = new URLSearchParams({
+          origin: `${startPoint.lng},${startPoint.lat}`,
+          destination: `${placeData.position.lng},${placeData.position.lat}`,
+        });
+
+        const requestUrl = `${BASED_URL.KAKKAO_DIRECTION}?${queryParams}`;
+
+        try {
+          const response = await instance.get(requestUrl, {
+            headers: {
+              Authorization: `KakaoAK ${process.env.NEXT_PUBLIC_CLIENT_ID_KAKAO_REST}`,
+            },
+          });
+
+          const elapsedTime = response.data.routes[0].summary.duration;
+          const path = extractPath(response.data);
+          setPolylinePath(path);
+          setDuration(elapsedTime);
+        } catch (error) {
+          setPolylinePath([]);
+          setDuration(0);
+        } finally {
+          setIsLoadingDirection(false);
+        }
       }
-      if (searchCoordinate !== null) {
-        setValidKeyword(true);
-        setStartPoint(searchCoordinate);
-      }
+    };
 
-      setShowMessage(true);
+    if (isValidCoordinate) {
+      getDestinationDirection();
     }
-  };
-
-  const handleChangeStartingPoint = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const location = event.target.value;
-    setShowMessage(false);
-    setCustomLocation(location);
-  };
-
-  const handleSelectLocation = (location: string) => {
-    setCustomLocation(location);
-    setRelatedLocation([]);
-
-    setTimeout(() => {
-      handleStartingPoint();
-    }, 300);
-  };
+  }, [isValidCoordinate, startPoint]);
 
   return (
     <section className="flex flex-col justify-center items-center gap-30">
@@ -108,14 +84,9 @@ const ProductDescription = () => {
           ) : (
             <CustomLocation
               destinationName={placeData.title}
-              customLocation={customLocation}
               elapsedTime={convertTime(duration)}
-              changeStartingPoint={(changelocation) => handleChangeStartingPoint(changelocation)}
-              customStartingPoint={() => handleStartingPoint()}
-              showMessage={showMessage}
-              validKeyword={validKeyword}
-              relatedLocation={relatedLocation}
-              handleSelectLocation={handleSelectLocation}
+              setStartPoint={setStartPoint}
+              isLoadingDirection={isLoadingDirection}
             />
           )}
           <ProductMap
